v2.0 beta

+++ bot/bypasser/bypass.py
@@ -1,34 +1,55 @@
-from bot import app
+from bot import app, logger
 from pyrogram import filters
 import PyBypass as bypasser
 
 
 @app.on_message(filters.command('bypass'))
-def bypass(_, msg):
-    if msg.reply_to_message:
-        url = msg.reply_to_message.text
-        if len(msg.text.split())>1:
-            nam = msg.text.split()[1]
-        else:
-            nam = None
-    else:
-        try:
-            url = msg.text.split()[1]
-            try:
-                nam = msg.text.split()[2]
-            except:
-                nam=None
-        except:
-            return msg.reply_text("Please Reply to a Url")
-
-    x = msg.reply(f"Bypassing __`{url}`__...")
+async def bypass(_, msg):
+    m = msg.text.split()
+    
+    if len(m)<2:
+        await msg.reply_text("Format: /bypass <url> <name only if required>")
+        return
+    if len(m)>2:
+        nam = m[2]
+        
+    url = m[1]
+    x = await msg.reply_text(f"Trying to Bypass __`{url}`__...")
 
     try:
-        bypassed = bypasser.bypass(url)
         if nam:
             bypassed = bypasser.bypass(url, name=nam)
+        bypassed = bypasser.bypass(url)
     except Exception as e:
-        return msg.reply(e)
+        x.delete()
+        await msg.reply_text(f"Couldn't bypass {url}\nError: {e}", disable_webpage_preview=True)
+        return
 
-    x.delete()
-    msg.reply_text(f"**BYPASSED URL:** `{bypassed}`")
+    await x.delete()
+    await msg.reply_text(f"**BYPASSED URL:** `{bypassed}`")
+
+
+@app.on_message(filters.command('gdtot'))
+def gdtotbypass(_, msg):
+    m = msg.text.split()
+    
+    if len(m)<3:
+        await msg.reply_text("Format: /bypass <url> <gdtot crypt>")
+        return
+    
+    url = m[1]
+    crypt = m[2]
+    
+    x = await msg.reply_text(f"Trying to Bypass __`{url}`__...")
+    
+    try:
+        bypassed = bypasser.bypass(url, gdtot_crypt=crypt)
+    except Exception as e:
+        await x.delete()
+        await msg.reply_text(f"Couldn't bypass {url}\nError: {e}", disable_webpage_preview=True)
+        return
+
+    await x.delete()
+    await msg.reply_text(f"**BYPASSED URL:** `{bypassed}`")
+    return
+
